fix(tracing) Omit the top level status tag
In the product side we don't really want people to search by this tag as
it is far more expensive than the `transaction.status` property which is
indexed separately. By not emitting this tag and only including it in
the trace context we won't end up with poor performing tags for users to
click on.