fix(tracing) Omit the top level status tag (#644)
In the product side we don't really want people to search by this tag as
it is far more expensive than the `transaction.status` property which is
indexed separately. By not emitting this tag and only including it in
the trace context we won't end up with poor performing tags for users to
click on.

To workaround get_trace_context() being called multiple times I needed
an additional non-tags place to store the status. I've had to shim the
status back into tags as non-transaction spans expect to have status as
a tag.

Co-authored-by: Markus Unterwaditzer <markus@unterwaditzer.net>