ref: Remove Hub.current is not None checks
By construction, Hub.current is never None, such that the expression

    Hub.current is not None

always evaluates to True.

This commit simplifies all uses of Hub.current, and in particular
chooses to write "return Hub.current.method(...)" for every method, even
when the method returns None. The intent is to make it easier to keep
the static API matching the Hub behavior. Without this, if a method
returns anything other than None the static API would silently drop it,
leading to unnecessary debugging time spent trying to identify the
culprit.