ref: Store tracked spans on start not finish
This matches the JS implementation. Without it, we cannot use the span
recorder of a span to find its parent transaction.

Note about test changes

Instrumented subprocess methods are called in this order: __init__,
communicate, wait. Because we now store the spans on start, that's the
order we expect the spans to be in. The previous order was based on
finish time.

Grouping the assertion of "op" values together produces better output on
failure, because one can easily detect what all the "op" values are,
instead of being left with only the first one that is different.

Similar to subprocess changes, the order of expected middleware spans in
Django is now sorted by start time.