fix: Re-add Auth.store_api_url
While we don't consider Auth to be part of the public API, the contract
is not very clear. This change is to prevent unnecessarily breaking
downstream uses of the Auth.store_api_url property.

Since we don't have any existing use of Python's DeprecationWarning, nor
any other system in place to communicate deprecation, we start with just
a note in the docstring.