feat: Avoid truncating span descriptions
For database auto-instrumented spans, the description contains
potentially long SQL queries that are most useful when not truncated.

Because arbitrarily large events may be discarded by the server as a
protection mechanism, we dynamically limit the description length,
preserving the most important descriptions/queries.

Performance impact

Preliminary CPU profiling using [1] suggests that uuid4() dominates the
execution time for code sending many transactions sequentially.

Preliminary memory profiling using [2] and looking at the max RSS of a
benchmark script suggests that the max RSS has no significant change
(JSON encoding in CPython is implemented in C).

In any case, we mitigate any increase in memory usage and run time for
the majority of cases by avoiding any extra work when the total number
of bytes consumed by descriptions do not exceed ~512 KB, which is
equivalent to having the standard string truncation applied.

Integrating profiling to the SDK is left for a future PR.

[1]: https://pypi.org/project/zprofile/
[2]: /usr/bin/time -l (macOS)