Travis CI: Test on Python 3.9 release candidate 1
Xenial is already the default dist on Travis CI.  (bionic and focal are also available)

In allow_failures mode until the GA release in one month.  https://devguide.python.org/#status-of-python-branches

Note: Python 3.4 is end of life and 3.5 will be end of life in two weeks.  https://devguide.python.org/devcycle/#end-of-life-branches

Fixes: