fix(aws): Don't crash if `event` isn't a single dict (#915)
Per https://docs.aws.amazon.com/lambda/latest/dg/python-handler.html, the `event` argument passed to the lambda function handler can be any jsonifiable type - string, int, list, etc - rather than just the dictionary we've previously assumed it to be. (This is particularly relevant for batch requests, which come in as a list of event dictionaries.)

When faced with such an `event`, our current integration crashes, because it tries to run `.get()` on it. This fixes that, by introducing the following behavior:

- If `event` is a list, tag the transaction as a batch and with the batch size.
- If `event` is a list, take the first entry as representative for the purposes of grabbing request data.
- If `event` (or the representative) isn't a dictionary, handle it gracefully and move on without request data.