Add build scripts for building and publishing AWS Lambda layers.
These are my scripts that I wrote to package a Sentry SDK for AWS Lambda.

Lambda layers are being uploaded as zip archives, then being extracted to
`/opt` directory in Lambda runtime.
And if you put your package to somewhere like `/opt/python/lib/python3.8/site-packages`
then it will be available to import from lambda function code.

Having such layers that users can attach to their lambda function is nice because
it's easier to integrate, especially in cases when lambda function does
not have any required packages except Sentry SDK.

So i wrote a couple of scripts.
One that builds the zip bundle with a specific file structure for each Python
version, e.g. `python/lib/python3.8/site-packages/sentry_sdk/`,
`python/lib/python3.7/site-packages/sentry_sdk` and so on.

And the one that publishes a layer to a set of AWS regions. It's important to publish
layer to every region the user might want because layer region and
lambda function region must match.

To run the publish script, it's required to set `AWS_ACCESS_KEY_ID` and
`AWS_SECRET_ACCESS_KEY` environment variables that `boto3` library
understands.

For each Python version, publish-script prints a list of obtained ARNs.
Having an ARN for the region, user could just paste it in the settings
and the package will be available to the runtime.

So it's necessary to put the list with latest ARNs somewhere in
documentation.