test(rq): Change test for enqueued_at and started_at
The started_at attribute of a job becomes None after the worker
finishes, because the worker calls the cleanup method of the job,
which erases some of the meta-information. So immediately after
creating a job, we call save to save this information to redis
and after refresh to load it. The same technique is used in the rq
library itself when testing the job module.