feat(transport): Client Report Support (#1181)
This adds support for client reports to the python SDK. This will cause the SDK
to send a report once every 30 seconds or once a minute. After 30 seconds it
will attempt to attach the report to a scheduled envelope if there is one, after
60 seconds it will send it as a separate envelope. Attempts of sending are
only made as a byproduct of attempted event / envelope sending or an
explicit flush.