Update Flask and Quart integrations (#1520)
Flask and Quart are deprecating and removing the ``_xxx_ctx_stack``s
and adopting a more direct usage of ContextVars. The previous code
will therefore break for the latest version of Quart and start to warn
for Flask and then break.

This fix should work with any version of Flask or Quart, and hence is
a more robust version. There is an extra indirection, however I don't
think this is on any hot path.

Co-authored-by: Anton Pirker <anton.pirker@sentry.io>