feat(profiling): Add support for profiles_sample_rate (#1613)
This changes the way profiling is enabled in the python sdk by allowing the end
user to specify a `profiles_sample_rate` which is used to control the sampling
of profiles. This sample rate is relative to the `traces_sample_rate` meaning
the true sample rate of profiles is approximately equal to
`traces_sample_rate * profiles_sample_rate`.