fix(profiling): Dynamically adjust profiler sleep time (#1634)
Because more time may have elapsed between 2 samples due to us calling the
sampling function and other threads executing, we need to account for it in the
sleep or the time between samples will often be greater than the expected
interval. This change ensures we account for this time elapsed and dynamically
adjust the amount of time we sleep for between samples.