ref(perf-issues): Increase max string size for desc
Our python sdk is the only sdk which sends truncated desc from the sdk side. This effects our ability to cleanly detect perf issues, but in general we should probably aim for more consistency. This bumps the max limit by a moderate amount (again, other sdk's are already sending unbounded data). We could potentially go higher (`2048`?). Partially put this up for RFC, but if this is fine then I'll fix tests where required and merge it in.