fix(profiling): Resolve inherited method class names
Methods may be inherited from a parent class. If multiple classes inherit from
the same class and uses the inherited method, we'd want it to report the parent
class's name instead of the individual child classes since they'd have the same
filename and lineno of the parent class and not the children.