feat(profiling): Introduce active thread id on scope
Upto this point, simply taking the current thread when the transaction/profile
was started was good enough. When using ASGI apps with non async handlers, the
request is received on the main thread. This is also where the transaction or
profile was started. However, the request is handled on another thread using a
thread pool. To support this use case, we want to be able to set the active
thread id on the scope where we can read it when we need it to allow the active
thread id to be set elsewhere.