fix(serializer): Add support for `byterray` and `memoryview` built-in types
Both `bytearray` and `memoryview` built-in types weren't explicitly
mentioned in the serializer logic, and as they are subtyping Sequence,
this led their instances to be enumerated upon, and to be output as a
list of bytes, byte per byte.

In the case of `memoryview`, this could also lead to a segmentation
fault if the memory referenced was already freed and unavailable to the
process by then.

By explicitly adding them as seralizable types, bytearray will be
decoded as a string just like bytes, and memoryview will use its
__repr__ method instead.

Close GH-1829