fix(profiling): Defaul in_app decision to None
Currently, the SDK marks all frames as in_app when it can't find any in_app
frames. As we try to move some of this detection server side, we still want to
allow the end user to overwrite the decision client side. So we'll leave in_app
as `None` to indicate the server should decide of the frame is in_app.