ref(profiling): Use the transaction timestamps to anchor the profile
We want the profile to be as closely aligned with the transaction's timestamps
as possible to make aligning the two visualizations as accurate as possible.
Here we change the transaction's internal `_start_timestamp_monotonic` to
contain an unit for each of the possible clocks we use in the various python
versions. This allows us to use the `start_timestamp` of the transaction as the
timestamp of the profile, and we can use the `_start_timestamp_monontonic` as
the anchor for all the relative timestamps in the profile.