fix(profiling): Handle non frame types in profiler
We've received report that occasionally, there's `AttributeError` on `f_back`.
It's unclear what exactly causes this issue because the source of the frame is
from a system libray. This avoids the `AttributeError` by wrapping the line in
question with a `try ... except ...`. And whenever it does encounter this error,
we should continue with what frames we have.