Add `loguru` integration
Actually, this is the solution in comments under #653 adapted to
codebase and tested as well.
https://github.com/getsentry/sentry-python/issues/653#issuecomment-788854865

I also changed `logging` integration to use methods instead of
functions in handlers, as in that way we can easily overwrite parts
that are different in `loguru` integration. It shouldn't be a problem,
as those methods are private and used only in that file.

Please let me know if I missed something, or you have more ideas how to
test this.