fix(profiling): Do not keep reference to frame to prevent memory leak (#2049)
The profiler can capture frames from it's own thread. When it does so, it holds
on to a reference to the frame in the previous sample. One of the frames it
holds on it is a frame from the profiler itself, which prevents the references
to other frames to other frames from being freed. A consequence of this is that
the local variables of those frames are not able to be freed either. This change
ensures we do not keep a reference to the profiler around in order to prevent
this issue.