Work with a copy of request, vars in the event
In some cases we were attaching parts of the original
request to the event with live references in them and
ending up modifying the underlying headers or request
data when we scrubbed the event. Now we make sure to
only attach a copy of the request to the event. We also
do the same for frame vars.