feat(starlette): add graphql.operation_name tag
Set a custom tag 'graphql.operation_name' if operation_name is
present in the JSON body or as a query parameter. This allows
GraphQL operations to be search for and alerted on. Previously,
no searchable tag existed because all GraphQL queries/mutations
are typically served from the same route and url.