fix(profiler): Add function name to profiler frame cache
Wrapper functions can take on the same name as the wrapped function. This means
that if a decorator is used to wrap different functions, even though the
filename and line number will be the same for all instances of the frame, the
function name can vary. Add the function name to the cache to avoid these cache
collisions.