Fix trace context in event payload (#2205)
Make sure that always a trace context is added to the event payload. But also make sure that if there is already a trace context in the event, do not overwrite it. (This used to be the behavior before tracing without performance. See: https://github.com/getsentry/sentry-python/blob/1.25.1/sentry_sdk/scope.py#L420)

---------

Co-authored-by: Ivana Kellyerova <ivana.kellyerova@sentry.io>