[Hackweek] Add explain plan to db spans. (#2315)
This is a proof of concept of adding the explain plan to db spans. The explain plan will be added to the span in the `db.explain_plan` data item. 

There is a cache to make sure that the explain plan for each db query is only executed ever X seconds and there is also a max number of elements that are cached. To make sure we do not put to much strain on CPU or memory.

Usage:
```
sentry_sdk.init(
    dsn="...",
    _experiments={
        "attach_explain_plans": {
            "explain_cache_size": 1000,  # Run explain plan for the 1000 most run queries
            "explain_cache_timeout_seconds": 60 * 60 * 24,  # Run the explain plan for each statement only every 24 hours
            "use_explain_analyze": True,  # Run "explain analyze" instead of only "explain"
        }
    }
```

Now you have a explain in the `span.data.db.explain_plan` in your database spans. 

---------

Co-authored-by: Ivana Kellyerova <ivana.kellyerova@sentry.io>