refactor: gRPC code generation script to be run from project root
I am not sure if the previous version was on purpose only running
from inside the test directory, however, I think this is more convenient
because the project root is the usual PWD during development.

Furthermore, I append the test directory to sys.path so we don't need
to tamper with the gRPC autogenerated code for the imports to work.