Load AWS Lambda secrets in Github CI (#2153)
Make sure our AWS Lambda test setup is correct and the tests work as expected and also in a timely manner.

We run our test in AWS Lambda and then parse the log output to see what events/envelopes where sent. Because Lambda truncates this log output to 4kb, I had to change the tests to make the events/envelopes smaller in size to get the whole event/envelop in the log output.

When the AWS env vars where not set, the tests where skipped but it looked like they where successful. I made them now fail loudly in that case, so we see if they do not run.

Also made the code easier to comprehend.

---------

Co-authored-by: Ivana Kellyerova <ivana.kellyerova@sentry.io>