feat(integrations): Support Django 5.0 (#2490)
Fix the way we wrap signal receivers: Django 5.0 introduced async receivers and changed the signature of the `Signal._live_receivers` method to return both sync and async receivers.

We'll need to change the Django version in tox.ini to 5.0 once it's been released. At the moment we're using the 5.0b1 release.