Do not create a span when task is triggered by Celery Beat (#2510)
We create a span for submitting a Celery task for execution (when apply_async() is called). In cases where web frameworks are calling apply_async() this is fine, because the web framework created a transaction where the span is attached.

When Celery Beat wakes up and is calling apply_async() this is not good, because there is no transaction and then the span ID of the newly created span will be given to the task as parent_span_id leading to orphaned transactions.

So in case apply_async() is called by Celery Beat, we do not create a span for submitting the task for execution.