feat: Code locations for metrics (#2526)
DDM wants to show code locations with metrics. Locations are semi-static information: they change infrequently, so they don't need to be reported with every data point.

Sentry expects locations to be reported at least once per day. With backdating of metrics, the timestamp used to report the location is the metric bucket's timestamp rounded down to the start of the day (UTC timezone).

The metrics aggregator keeps a cache of previously reported locations. When a location is seen for the first time on a day, it is added to a list of pending locations. On the next flush cycle, all pending locations are sent to Sentry in the same envelope as the metric buckets.

See: https://github.com/getsentry/relay/pull/2751
Epic: https://github.com/getsentry/sentry/issues/60260
---------

Co-authored-by: Armin Ronacher <armin.ronacher@active-4.com>
Co-authored-by: Anton Pirker <anton.pirker@sentry.io>