fix(integrations): Fix Lambda integration with EventBridge source
When a lambda is triggered by an AWS EventBridge pipe the record
contains an explicit "header" key with an empty list. This breaks the
assumption that headers is always a dict or None. Update the
AwsLambdaIntegration to explicitly verify that header is a dict before
passing it on to the `continue_trace` function.

Fixes GH-2545