Hash AWS Lambda test functions based on current revision (#2557)
We were using the current SDK version for determining whether an AWS Lambda function should be reused, so e.g. on PRs, this would reuse the existing functions instead of creating new ones with any changes from the PR. Changing this to use the current commit instead.

Also, use a 6 character hash instead of 5 characters, just to lower the chance for collisions a bit.