fix(ci): Fix yaml generation script (#2695)
The generation script was supposed to raise an error if it detected that the integration test YAML files have changed but haven't been committed. The check is based on a hash of the contents of the YAML files, but there was a typo in the file names to consider (`integration` -> `integrations`), so it wasn't actually looking at any files and was always trivially true.

Now it'll properly complain if changes are made to `tox.ini` or to some of the constants in the splitting script that result in new YAML files, but those are not part of the commit.