test: Fix `pytest` error (#2712)
The ability to pass None to pytest.capture_warnings was removed in pytest version 8.0.0. To validate that this fix, one can run any of the test cases with pytest==8.0.0. Without this change, the test immediately fails with an error; with the change, the test suite runs as expected.

Fixes GH-2693