fix(query-source): Fix query source relative filepath (#2717)
When generating the filename attribute for stack trace frames, the SDK uses the `filename_for_module` function. When generating the `code.filepath` attribute for query spans, the SDK does not use that function. Because of this inconsistency, code mappings that work with stack frames sometimes don't work with queries that come from the same files.

This change makes sure that query sources use `filename_for_module`, so the paths are consistent.