Scope refactoring (merge Hubs and Scopes) (#2610)
This refactors the SDK to move away from the Hub and have all the functionality in the Scope. Introducing different types of scopes. This aligns the SDK with how Opentelementry (OTel) handles data bringing us closer to be 100% OTel compatible.

This change was discussed in this RFC: 
https://github.com/getsentry/rfcs/pull/122

There is also a small FAQ: 
https://gist.github.com/mitsuhiko/1bc78d04ea7d08e5b50d27e42676db80

And a Miro board showing how the new scopes manage data: 
https://miro.com/app/board/uXjVNtPiOfI=/?share_link_id=216270218892

### This RP contains
- Introduction of global, isolation, and current scope
- Deprecation of the Hub
- All existing Hub based API still works and is still used by most of our integrations. Under the hood the new Scopes are used.
- (this PR now includes all the changes made in the [first PR](https://github.com/getsentry/sentry-python/pull/2609) introducing the new API)

### Breaking changes
- The Pyramid integration will not capture errors that might happen in `authenticated_userid()` in a custom `AuthenticationPolicy` class.
- The parameter `propagate_hub` in `ThreadingIntegration()` was deprecated and renamed to `propagate_scope`.