ref(api): Abstract base classes (#2667)
We have some classes in the SDK that essentially function as abstract classes, since they have some methods that raise NotImplementedError when called, since they are intended to be overriden by a subclass. Now that all supported Python versions support abstract base classes (ABCs), we should refactor these classes to be ABCs. Making this change will explicitly indicate which methods need to be overridden by a subclass.

Changing a public class into an ABC is a breaking change, since instantiating the class, or any subclass that hasn't overridden all abstract methods, will raise an error. Therefore, I have added this (draft) PR to the SDK 2.0 milestone.

* Convert `Transport` class into an ABC

* ABC metrics

* ABC scheduler

* ABC integration

* RequestExtractor comment

* Deprecate `stop_profiling` and stop calling it

This change is required because otherwise, the linter complains about the `stop_profiling` being an empty concrete method in an abstract class.

* Actually, let's remove `stop_profiling`

* Add ABCs to migration guide

* fix mypy

* Make docstring more relevant to users

* `ensure_running` no longer abstract

* Fix mypy