docs: Move transaction __init__ doc comment content (#2918)
This change moves the Transaction constructor's parameter doctsring from the __init__ method to the class's docstring. This way, the API docs display the parameter descriptions under the class. When the docstring is defined on __init__, the parameter descriptions are missing from the API docs.

This change also documents the kwargs parameter in the API docs.

ref https://github.com/getsentry/sentry-docs/issues/5082