docs: Document top-level `start_transaction` function (#2924)
Although the @scopemethod decorator should take care of adding this documentation comment to start_transaction, the decorator does not appear to work when hovering over the sentry_sdk.start_transaction function in VSCode. Adding the doc comment explicitly allows the documentation to be shown when hovering in VSCode, but it does not change the output of our API docs, which still uses the @scopemethod output.

This is the final PR in this repository that addresses getsentry/sentry-docs#5082; the last step to closing that issue will be to add a link in the docs to start_transaction in our API docs. We can only do this once 2.0 is released.