Add Lambda function that deletes test Lambda functions (#2960)
Lambda function that deletes all test lambda functions (prefixed with `test_`) that have been created during CI runs. * Lambda function that deletes test Lambda functions. This function is run every Sunday and it is monitored in the `sentry-python` project on Sentry.io with the Crons feature and it is also emitting metrics on how many functions it deletes in each run. 
---------

Co-authored-by: Ivana Kellyerova <ivana.kellyerova@sentry.io>