fix(tracing): Only propagate headers from spans within transactions
This change ensures that we only propagate trace headers from spans that are within a transaction. This fixes a bug where any child transactions of a span created outside a transaction are missing a dynamic sampling context and are part of a trace missing a root transaction (because the root is the span).

Also, remove/modify tests that were asserting the old behavior.

Fixes #3068