fix(django): Proper transaction names for i18n routes (#3104)
`pattern.pattern._route` for i18n'd Django routes is a proxy object rather than a string. This causes an exception in the resolver, leading to the transaction not getting a proper name but rather falling back to the default `Generic WSGI request`.

The string representation of the proxy object is the actual desired endpoint route, so let's use that.