docs(init): Fix `sentry_sdk.init` type hint
The current type hint suggests that all the parameters can be passed as positional arguments, when this is not the case. Only the `dsn` can be passed as a positional argument; the rest must be passed as keyword arguments.

This PR makes the type hint reflect the reality of what parameters can be passed to `sentry_sdk.init`.